+++ lib/src/jikan.dart
@@ -14,58 +14,49 @@ import 'package:jikan_api/src/model/club/club.dart';
 import 'package:jikan_api/src/model/club/member.dart';
 import 'package:jikan_api/src/model/common/article.dart';
 import 'package:jikan_api/src/model/common/forum.dart';
-import 'package:jikan_api/src/model/common/more_info.dart';
 import 'package:jikan_api/src/model/common/picture.dart';
 import 'package:jikan_api/src/model/common/recommendation.dart';
 import 'package:jikan_api/src/model/common/review.dart';
 import 'package:jikan_api/src/model/common/stats.dart';
 import 'package:jikan_api/src/model/common/user_update.dart';
+import 'package:jikan_api/src/model/constants.dart';
 import 'package:jikan_api/src/model/genre/genre.dart';
 import 'package:jikan_api/src/model/genre/genre_list.dart';
-import 'package:jikan_api/src/model/genre/genre_type.dart';
 import 'package:jikan_api/src/model/magazine/magazine.dart';
 import 'package:jikan_api/src/model/manga/manga.dart';
 import 'package:jikan_api/src/model/person/person.dart';
 import 'package:jikan_api/src/model/producer/producer.dart';
 import 'package:jikan_api/src/model/schedule/schedule.dart';
-import 'package:jikan_api/src/model/schedule/week_day.dart';
 import 'package:jikan_api/src/model/search/search.dart';
-import 'package:jikan_api/src/model/search/search_type.dart';
 import 'package:jikan_api/src/model/season/season.dart';
 import 'package:jikan_api/src/model/season/season_archive.dart';
-import 'package:jikan_api/src/model/season/season_type.dart';
 import 'package:jikan_api/src/model/serializers.dart';
 import 'package:jikan_api/src/model/top/top.dart';
-import 'package:jikan_api/src/model/top/top_type.dart';
 import 'package:jikan_api/src/model/user/friend.dart';
 import 'package:jikan_api/src/model/user/history.dart';
-import 'package:jikan_api/src/model/user/history_type.dart';
 import 'package:jikan_api/src/model/user/user_item.dart';
 import 'package:jikan_api/src/model/user/user_profile.dart';
-import 'package:jikan_api/src/model/user/user_request_type.dart';
 
 class Jikan {
-  final String baseUrl = 'https://api.jikan.moe/v3';
-
   Future<String> _getResponse(String url) async {
     http.Response response;
-    print(url);
+    print(baseUrl + url);
     do {
-      response = await http.get(url);
+      response = await http.get(baseUrl + url);
     } while (response.statusCode == 429 || response.statusCode == 500);
 
     return response.body;
   }
 
   Future<Anime> getAnimeInfo(int animeId) async {
-    var url = baseUrl + '/anime/$animeId';
+    var url = '/anime/$animeId';
     var response = await _getResponse(url);
 
     return Anime.fromJson(response);
   }
 
   Future<CharacterStaff> getAnimeCharactersStaff(int animeId) async {
-    var url = baseUrl + '/anime/$animeId/characters_staff';
+    var url = '/anime/$animeId/characters_staff';
     var response = await _getResponse(url);
 
     return CharacterStaff.fromJson(response);
@@ -73,7 +64,7 @@ class Jikan {
 
   Future<BuiltList<Episode>> getAnimeEpisodes(int animeId,
       {int page = 1}) async {
-    var url = baseUrl + '/anime/$animeId/episodes/$page';
+    var url = '/anime/$animeId/episodes/$page';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -83,7 +74,7 @@ class Jikan {
   }
 
   Future<BuiltList<Article>> getAnimeNews(int animeId) async {
-    var url = baseUrl + '/anime/$animeId/news';
+    var url = '/anime/$animeId/news';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -93,7 +84,7 @@ class Jikan {
   }
 
   Future<BuiltList<Picture>> getAnimePictures(int animeId) async {
-    var url = baseUrl + '/anime/$animeId/pictures';
+    var url = '/anime/$animeId/pictures';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -103,7 +94,7 @@ class Jikan {
   }
 
   Future<BuiltList<Promo>> getAnimeVideos(int animeId) async {
-    var url = baseUrl + '/anime/$animeId/videos';
+    var url = '/anime/$animeId/videos';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -113,14 +104,17 @@ class Jikan {
   }
 
   Future<Stats> getAnimeStats(int animeId) async {
-    var url = baseUrl + '/anime/$animeId/stats';
+    var url = '/anime/$animeId/stats';
     var response = await _getResponse(url);
 
     return Stats.fromJson(response);
   }
 
-  Future<BuiltList<Forum>> getAnimeForum(int animeId) async {
-    var url = baseUrl + '/anime/$animeId/forum';
+  Future<BuiltList<Forum>> getAnimeForum(int animeId, {ForumType type}) async {
+    var url = '/anime/$animeId/forum';
+    if (type != null) {
+      url += '/${type.toString().split('.')[1]}';
+    }
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -129,15 +123,16 @@ class Jikan {
     return serializers.deserialize(topics, specifiedType: listForum);
   }
 
-  Future<MoreInfo> getAnimeMoreInfo(int animeId) async {
-    var url = baseUrl + '/anime/$animeId/moreinfo';
+  Future<String> getAnimeMoreInfo(int animeId) async {
+    var url = '/anime/$animeId/moreinfo';
     var response = await _getResponse(url);
 
-    return MoreInfo.fromJson(response);
+    var jsonEncoded = json.decode(response);
+    return jsonEncoded['moreinfo'];
   }
 
   Future<BuiltList<Review>> getAnimeReviews(int animeId, {int page = 1}) async {
-    var url = baseUrl + '/anime/$animeId/reviews/$page';
+    var url = '/anime/$animeId/reviews/$page';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -147,7 +142,7 @@ class Jikan {
   }
 
   Future<BuiltList<Recommendation>> getAnimeRecommendations(int animeId) async {
-    var url = baseUrl + '/anime/$animeId/recommendations';
+    var url = '/anime/$animeId/recommendations';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -159,7 +154,7 @@ class Jikan {
 
   Future<BuiltList<UserUpdate>> getAnimeUserUpdates(int animeId,
       {int page = 1}) async {
-    var url = baseUrl + '/anime/$animeId/userupdates/$page';
+    var url = '/anime/$animeId/userupdates/$page';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -169,14 +164,14 @@ class Jikan {
   }
 
   Future<Manga> getMangaInfo(int mangaId) async {
-    var url = baseUrl + '/manga/$mangaId';
+    var url = '/manga/$mangaId';
     var response = await _getResponse(url);
 
     return Manga.fromJson(response);
   }
 
   Future<BuiltList<CharacterRole>> getMangaCharacters(int mangaId) async {
-    var url = baseUrl + '/manga/$mangaId/characters';
+    var url = '/manga/$mangaId/characters';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -186,7 +181,7 @@ class Jikan {
   }
 
   Future<BuiltList<Article>> getMangaNews(int mangaId) async {
-    var url = baseUrl + '/manga/$mangaId/news';
+    var url = '/manga/$mangaId/news';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -196,7 +191,7 @@ class Jikan {
   }
 
   Future<BuiltList<Picture>> getMangaPictures(int mangaId) async {
-    var url = baseUrl + '/manga/$mangaId/pictures';
+    var url = '/manga/$mangaId/pictures';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -206,14 +201,17 @@ class Jikan {
   }
 
   Future<Stats> getMangaStats(int mangaId) async {
-    var url = baseUrl + '/manga/$mangaId/stats';
+    var url = '/manga/$mangaId/stats';
     var response = await _getResponse(url);
 
     return Stats.fromJson(response);
   }
 
-  Future<BuiltList<Forum>> getMangaForum(int mangaId) async {
-    var url = baseUrl + '/manga/$mangaId/forum';
+  Future<BuiltList<Forum>> getMangaForum(int mangaId, {ForumType type}) async {
+    var url = '/manga/$mangaId/forum';
+    if (type != null) {
+      url += '/${type.toString().split('.')[1]}';
+    }
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -222,15 +220,16 @@ class Jikan {
     return serializers.deserialize(topics, specifiedType: listForum);
   }
 
-  Future<MoreInfo> getMangaMoreInfo(int mangaId) async {
-    var url = baseUrl + '/manga/$mangaId/moreinfo';
+  Future<String> getMangaMoreInfo(int mangaId) async {
+    var url = '/manga/$mangaId/moreinfo';
     var response = await _getResponse(url);
 
-    return MoreInfo.fromJson(response);
+    var jsonEncoded = json.decode(response);
+    return jsonEncoded['moreinfo'];
   }
 
   Future<BuiltList<Review>> getMangaReviews(int mangaId, {int page = 1}) async {
-    var url = baseUrl + '/manga/$mangaId/reviews/$page';
+    var url = '/manga/$mangaId/reviews/$page';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -240,7 +239,7 @@ class Jikan {
   }
 
   Future<BuiltList<Recommendation>> getMangaRecommendations(int mangaId) async {
-    var url = baseUrl + '/manga/$mangaId/recommendations';
+    var url = '/manga/$mangaId/recommendations';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -252,7 +251,7 @@ class Jikan {
 
   Future<BuiltList<UserUpdate>> getMangaUserUpdates(int mangaId,
       {int page = 1}) async {
-    var url = baseUrl + '/manga/$mangaId/userupdates/$page';
+    var url = '/manga/$mangaId/userupdates/$page';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -262,14 +261,14 @@ class Jikan {
   }
 
   Future<Person> getPersonInfo(int personId) async {
-    var url = baseUrl + '/person/$personId';
+    var url = '/person/$personId';
     var response = await _getResponse(url);
 
     return Person.fromJson(response);
   }
 
   Future<BuiltList<Picture>> getPersonPictures(int personId) async {
-    var url = baseUrl + '/person/$personId/pictures';
+    var url = '/person/$personId/pictures';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -279,14 +278,14 @@ class Jikan {
   }
 
   Future<Character> getCharacterInfo(int characterId) async {
-    var url = baseUrl + '/character/$characterId';
+    var url = '/character/$characterId';
     var response = await _getResponse(url);
 
     return Character.fromJson(response);
   }
 
-  Future<BuiltList<Picture>> getCharactersPictures(int characterId) async {
-    var url = baseUrl + '/character/$characterId/pictures';
+  Future<BuiltList<Picture>> getCharacterPictures(int characterId) async {
+    var url = '/character/$characterId/pictures';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -297,7 +296,7 @@ class Jikan {
 
   Future<BuiltList<Search>> search(String query, SearchType type,
       {int page = 1}) async {
-    var url = baseUrl + '/search/${searchTypeString(type)}?q=$query&page=$page';
+    var url = '/search/${type.toString().split('.')[1]}?q=$query&page=$page';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -307,9 +306,9 @@ class Jikan {
   }
 
   Future<Season> getSeason({int year, SeasonType season}) async {
-    var url = baseUrl + '/season';
+    var url = '/season';
     if (year != null && season != null) {
-      url += '/$year/${seasonTypeString(season)}';
+      url += '/$year/${season.toString().split('.')[1]}';
     }
     var response = await _getResponse(url);
 
@@ -317,14 +316,14 @@ class Jikan {
   }
 
   Future<Season> getSeasonLater() async {
-    var url = baseUrl + '/season/later';
+    var url = '/season/later';
     var response = await _getResponse(url);
 
     return Season.fromJson(response);
   }
 
   Future<BuiltList<SeasonArchive>> getSeasonArchive() async {
-    var url = baseUrl + '/season/archive';
+    var url = '/season/archive';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -334,9 +333,9 @@ class Jikan {
   }
 
   Future<Schedule> getSchedule({WeekDay weekday}) async {
-    var url = baseUrl + '/schedule/';
+    var url = '/schedule';
     if (weekday != null) {
-      url += weekDayString(weekday);
+      url += '/${weekday.toString().split('.')[1]}';
     }
     var response = await _getResponse(url);
 
@@ -345,9 +344,9 @@ class Jikan {
 
   Future<BuiltList<Top>> getTop(TopType type,
       {TopSubtype subtype, int page = 1}) async {
-    var url = baseUrl + '/top/${topTypeString(type)}/$page';
+    var url = '/top/${type.toString().split('.')[1]}/$page';
     if (subtype != null) {
-      url += '/${topSubtypeString(subtype)}';
+      url += '/${subtype.toString().split('.')[1]}';
     }
     var response = await _getResponse(url);
 
@@ -359,36 +358,39 @@ class Jikan {
 
   Future<GenreList> getGenre(GenreType type, Genre genre,
       {int page = 1}) async {
-    var url = baseUrl + '/genre/${genreTypeString(type)}/${genre.value}/$page';
+    var url = '/genre/${type.toString().split('.')[1]}/${genre.value}/$page';
     var response = await _getResponse(url);
 
     return GenreList.fromJson(response);
   }
 
   Future<Producer> getProducerInfo(int producerId, {int page = 1}) async {
-    var url = baseUrl + '/producer/$producerId/$page';
+    var url = '/producer/$producerId/$page';
     var response = await _getResponse(url);
 
     return Producer.fromJson(response);
   }
 
   Future<Magazine> getMagazineInfo(int magazineId, {int page = 1}) async {
-    var url = baseUrl + '/magazine/$magazineId/$page';
+    var url = '/magazine/$magazineId/$page';
     var response = await _getResponse(url);
 
     return Magazine.fromJson(response);
   }
 
   Future<UserProfile> getUserProfile(String username) async {
-    var url = baseUrl + '/user/$username/profile';
+    var url = '/user/$username/profile';
     var response = await _getResponse(url);
 
     return UserProfile.fromJson(response);
   }
 
-  Future<BuiltList<History>> getUserHistory(
-      String username, HistoryType type) async {
-    var url = baseUrl + '/user/$username/history/${historyTypeString(type)}';
+  Future<BuiltList<History>> getUserHistory(String username,
+      {HistoryType type}) async {
+    var url = '/user/$username/history';
+    if (type != null) {
+      url += '/${type.toString().split('.')[1]}';
+    }
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -397,8 +399,9 @@ class Jikan {
     return serializers.deserialize(history, specifiedType: listHistory);
   }
 
-  Future<BuiltList<Friend>> getUserFriends(String username) async {
-    var url = baseUrl + '/user/$username/friends';
+  Future<BuiltList<Friend>> getUserFriends(String username,
+      {int page = 1}) async {
+    var url = '/user/$username/friends/$page';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);
@@ -407,10 +410,16 @@ class Jikan {
     return serializers.deserialize(friends, specifiedType: listFriend);
   }
 
-  Future<BuiltList<UserItem>> getUserAnimeList(String username, ListType type,
-      {String query, String order, String sort = 'desc', int page = 1}) async {
-    var url =
-        baseUrl + '/user/$username/animelist/${listTypeString(type)}/$page';
+  Future<BuiltList<UserItem>> getUserAnimeList(String username,
+      {ListType type,
+      String query,
+      String order,
+      String sort = 'desc',
+      int page = 1}) async {
+    var url = '/user/$username/animelist';
+    if (type != null) {
+      url += '/${type.toString().split('.')[1]}/$page';
+    }
     if (query != null) {
       url += '?q=$query';
     }
@@ -426,10 +435,16 @@ class Jikan {
     return serializers.deserialize(anime, specifiedType: listAnime);
   }
 
-  Future<BuiltList<UserItem>> getUserMangaList(String username, ListType type,
-      {String query, String order, String sort = 'desc', int page = 1}) async {
-    var url =
-        baseUrl + '/user/$username/mangalist/${listTypeString(type)}/$page';
+  Future<BuiltList<UserItem>> getUserMangaList(String username,
+      {ListType type,
+      String query,
+      String order,
+      String sort = 'desc',
+      int page = 1}) async {
+    var url = '/user/$username/mangalist';
+    if (type != null) {
+      url += '/${type.toString().split('.')[1]}/$page';
+    }
     if (query != null) {
       url += '?q=$query';
     }
@@ -446,14 +461,14 @@ class Jikan {
   }
 
   Future<Club> getClubInfo(int clubId) async {
-    var url = baseUrl + '/club/$clubId';
+    var url = '/club/$clubId';
     var response = await _getResponse(url);
 
     return Club.fromJson(response);
   }
 
   Future<BuiltList<Member>> getClubMembers(int clubId, {int page = 1}) async {
-    var url = baseUrl + '/club/$clubId/members/$page';
+    var url = '/club/$clubId/members/$page';
     var response = await _getResponse(url);
 
     var jsonEncoded = json.decode(response);