+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+## 1.3.0
+
+* Added club models.
+* Set optional param `type` to `getAnimeForum` and `getMangaForum`.
+* Set optional param `page` to `getUserFriends`.
+* `getAnimeMoreInfo` and `getMangaMoreInfo` now returns `String`.
+
 ## 1.2.0
 
-* Set "scoredBy" param to nullable.
+* Set `scoredBy` param to nullable.
 * Fixed linter warnings.
 
 ## 1.1.0